@@ -3,21 +3,21 @@
 
 #include "Utils.h"
 #include <iostream>
-#include <boost/property_tree/xml_parser.hpp>
 #include "PropertyParser.h"
 #include "ObjectParser.h"
 
 namespace CudaTracerLib {
 
-void ParseMitsubaScene(DynamicScene& scene, const std::string& scene_file, const std::map<std::string, std::string>& cmd_def_storage, boost::optional<Vec2i>& image_res, bool assume_rotated_coords, bool create_exterior_bssrdf, bool create_interior_bssrdf)
+void ParseMitsubaScene(DynamicScene& scene, const std::string& scene_file, const std::map<std::string, std::string>& cmd_def_storage, std::optional<Vec2i>& image_res, bool assume_rotated_coords, bool create_exterior_bssrdf, bool create_interior_bssrdf)
 {
-	ParserState S(scene, boost::filesystem::path(scene_file).parent_path().string(), assume_rotated_coords, create_exterior_bssrdf, create_interior_bssrdf);
+	ParserState S(scene, std::filesystem::path(scene_file).parent_path().string(), assume_rotated_coords, create_exterior_bssrdf, create_interior_bssrdf);
 	for (auto ent : cmd_def_storage)
 		S.def_storage.add(ent.first, ent.second);
 
-	boost::property_tree::ptree pt;
-	read_xml(scene_file, pt);
-	CudaTracerLib::XMLNode root("", pt);
+	pugi::xml_document doc;
+	if (!doc.load_file(scene_file.c_str()))
+		throw std::runtime_error("couldn't loader scene xml!");
+	CudaTracerLib::XMLNode root("", doc);
 
 	root.get_child_node("scene").iterate_child_nodes([&](const CudaTracerLib::XMLNode& n)
 	{
